CW-282: fix for location and range crumbs not showing
